fix ip retrieval. updated host ip

+++ src_server/check_carrier.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 CARRIER_PATH=/home/elabox/apps/ela.carrier
 BOOTSTRAP_FILE=${CARRIER_PATH}/bootstrapd.conf
-CURRENT_IP=$(curl -s ipinfo.io/ip)
+CURRENT_IP=$(curl -s ifconfig.me)
 echo "$CURRENT_IP"
 CARRIER_IP=$(grep "external_ip" ${BOOTSTRAP_FILE} | cut -d'"' -f 2)
 echo $CARRIER_IP
+
+if [ "$CURRENT_IP" == "" ]; then
+    echo "No IP found. Exiting..."
+    exit 1
+fi
 
 if [ "$CURRENT_IP" == "$CARRIER_IP" ]; then
     echo "Carrier is running fine"
@@ -14,6 +19,9 @@ else
     sed -i 's/external_ip = "'$CARRIER_IP'"/external_ip = "'$CURRENT_IP'"/g' ${BOOTSTRAP_FILE}
     # stop and restart the carrier
     CARRIER_PID=$(ps aux | grep [.]/ela-bootstrapd | head -1 | cut -d' ' -f 4)
-    kill ${CARRIER_PID}
+    echo "Killing carrier PID "${CARRIER_PID}...
+    if [ "$CARRIER_PID" != "" ]; then
+        kill ${CARRIER_PID}
+    fi
     ${CARRIER_PATH}/ela-bootstrapd --config=${BOOTSTRAP_FILE} --foreground
 fi
